+++ ricaun.Nuke.PackageBuilder/Components/Revit/IRevitPackageBuilder.cs
@@ -66,7 +66,7 @@ namespace ricaun.Nuke.Components
             // CopyInstallationFiles If Exists
             CopyInstallationFilesTo(PackageBuilderDirectory);
 
-            new RevitContentsBuilder(project, BundleDirectory, NewVersions)
+            new RevitContentsBuilder(project, BundleDirectory, MiddleVersions, NewVersions)
                 .Build(BundleDirectory / "PackageContents.xml");
 
             new IssRevitBuilder(project, PackageBuilderDirectory, IssConfiguration)