+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.3] / 2023-05-06
+### Updated
+- Update Readme
+- Update `RevitContentsBuilder` add `MiddleVersions`
+
 ## [1.4.2] / 2023-03-30
 ### Updated
 - Update `ricaun.Nuke` to `1.4.4`
@@ -224,6 +229,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.4.3]: ../../compare/1.4.2...1.4.3
 [1.4.2]: ../../compare/1.4.1...1.4.2
 [1.4.1]: ../../compare/1.4.0...1.4.1
 [1.4.0]: ../../compare/1.3.5...1.4.0